preparation for publication
