+++ Procfile
@@ -1,0 +1,1 @@
+bot: kotlin NewsBotApplication.kt
