+++ src/main/kotlin/ru/tinkoff/utils/InfoType.kt
@@ -1,4 +1,4 @@
-package ru.tinkoff
+package ru.tinkoff.utils
 
 enum class InfoType {
     NEWS, WEATHER, TRAFFIC
