--- src/main/resources/application.yml
@@ -1,3 +1,0 @@
-telegram:
-  botName: projectnewsbot
-  token: 5123491633:AAHDpSUrLfXSi8u1xG8XymCuPlyuGV0AGuA
